requirements.txt
aws-cdk-lib>=2.100.0
constructs
boto3
